Remove Statefullness from SchedulerEntry (#2940)
- Make SchedulerEntry, and therefore the schedulers, use polymorphism
instead of dispatching.
- Remote all stateful information from SchedulerEntry and therefore the
schedulers. Although it's stateless instances are still needed to use
the overloaded functions. So there's still dispatch in
`SchedulerEntry::makeSchedulerInstance` which dispatches on
`HeuristicType`, and makes a corresponding scheduler instance.
- Put all the stateful information in HeuristicParams.
- Removed the Scheduler's constructors, splitting out a separate public
call to `computeHeuristics`.
- Pass heuristics around through unique_ptr's for use with our
polymorphic type.
- Rename ScheduleHeuristic->HeuristicType.
- Remove some shallow "using" in fusion_segmenter.cpp.
- Rename `FusionHeuristics` to `HeuristicParamsList` and move it to
heuristics.h
- Renamed instances of HueristicParams to match the type of params when
known. i.e. `mparams` for `MatmulParams`, or `pparams` for
`PointwiseParams`. Just trying to make it consistent as we were mostly
using this convention, but not consistently.