Multi-matmul scheduler: add test and schedule smem operand store (#2913)
This PR follows up from #2719 toward the eventual goal of merging #2458.
Here I am introducing a parametrized test suite that I will use to check
that the generated code matches the current scheduler for a variety of
scenarios. I also am scheduling the first consumer tensor in the fusion:
the smem store of operands. This shows how we can selectively check that
the schedule is correct before we have everything implemented, since the
test is only checking these tensors and nothing else in the fusion.

In the next PRs, I will schedule the rest of the prologue, as well as
the mma result and the epilogue tensors. In each case, I will change the
test to check more and more of the fusion for correctness.

This PR introduces `AbstractMatmuTensor`, an `AbstractTensor` for which
each dimension can be tagged with a role. This lets us track the roles
of each dimension during abstract tensor scheduling. This is only
currently used in `blockTileTensors` and `mma_utils::makeTile`, but in
the future we can imagine using it more.