Multiple Matmul Scheduler: schedule prologue
This is the next PR in a series leading up to #2458. The previous PR 2913 scheduled the smem store and added the test framework.
This PR schedules the smem load and the prologue, i.e. everything up to
the mma instruction.

The next PRs in the series will do the following:
- Schedule the mma op result
- Schedule the epilogue including the split-K tensor
- Add test with horizontal fusion of matmuls (#2458)
- Switch to multiple matmul scheduler, i.e. remove old scheduleMatmul()
- Update matmul heuristic to account for horizontal fusion
- Update canSchedule checks to accept horizontally fused matmul segments