Extend MaxPosCalculartor::getMaxProducerPosFromConsumer to support non-conventional loop domains (#2983)
Part of #2902. See the linked design doc for the background context.

~Stacked on #2984~
Stacked on #2987 

The IdModel-based analysis is required for non-conventional loop domains
as the dependency from the logical to loop domains may not be just one
way from logical to loop. However, the IdModel-based analysis does not
support broadcast forwarding, so we can't just completely switch to the
new method.

For now, I think it makes most sense to use the existing method whenever
logical and loop domains are generated in the conventional scheduling
primitives such as split and merge. The new method is only used
otherwise.

I also considered implementing the broadcast forwarding in the
IdModel-based approach as well, but it doesn't seem to me worthwhile
doing so at this moment. I thought maybe using the Permissive graph
would make it trivial, but there seem to be (not a small number of)
subtle corner cases of mappings that may not be considered in the
Permissive graph, e.g., if the indexed domains should be mapped or not.

It should be certainly possible to extend the Permissive graph to
address these corner cases, but another factor I think we should
consider is that the broadcast forwarding itself may not be necessary
with the setLoopDomain-based approach. We should probably set the right
loop domain from the beginning instead of trying to make different loop
domains to be inlinable.