Support non-conventional loop domains in TensorView::updateMaxProducerPosition. (#3003)
Follow up to #2983.

`TensorView::updateMaxProducerPosition` is used to set the max producer
position when `inlineAt` is used. With this PR, `inlineMost` should work
with `setLoopDomain`.

This should be the last remaining piece to enable inlineMost with loop
domains set by `setLoopDomain`. There are a couple more issues to
address in lowering, though.