Multi-matmul scheduler: Skip device dims when counting local batch dims (#3065)
This patches a small bug that incorrectly counted dimensions. We use a
different system (`MatmulDimRolesMap`) to detect device and batch
dimensions now, and it was identifying some device dims as batch
dimensions. Although these dimensions _are_ batch dimensions in the
batch matmul sense, they are not _local_ batch dimensions, i.e. they are
not realized on each rank, so we skip them when computing
`num_local_batch_dims_`. By contrast, in the existing scheduler we
simply look for a device dim then subtract that plus 3 from
`mma_result->nDims()`, which is not as flexible for future supported use
cases with multiple M, N, or K dimensions.

This is necessary to get #3043 working.