Specify different-dtype alias TV alignment (#3084)
See https://github.com/NVIDIA/Fuser/pull/2934#discussion_r1783671457

PR #665 allowed us to re-use allocations that have different dtypes. We
already check that our aliased tensors do not have vectorized accesses
larger than those of the original tensors. However, when we have
different dtypes we `reinterpret_cast` it to a different `Array` type.
Previously we did not specify any alignment in that type's template
args, meaning it assumed an alignment of size 1. Since the actual
addresses will all still be aligned this does not caused misaligned
accesses at runtime. This PR sets the template arg for alignment to be
that of the vectorized access width for the alias tensor, so that the
compiler could hypothetically do some optimizations knowing the address
is aligned.