Preliminary support of loop graph and "promotion" with setLoopDomain
I'm still experimenting with IdModel setLoopDomain, but this PR at least
allows for correct loop graph and promotion maps to be generated as long
as no forwarding-like scheduling is involved in loop domains that are
not fully derived from logical domains.

The existing slice tests with setLoopDomain are extended with inlineMost
since it's now fully working.

Note that there is still an issue of inlining analysis (#3068).
Addressing it is not in the scope of this PR.