Generate loops based on the IdModel Loop graph (promotion) (#3036)
This PR introduces an option to use loop promotion IDs when generating
`ForLoop` IR nodes. Consequently, loop index variables are also
maintained by `IdModel`, not by `ComputeAtMap`.

This is required when loop domains are not directly derived from logical
domains as the ComputeAtMap loop map does not support such loop domains.

For now, I decided to add yet another IdModel option
(`IdModelEnableOption::Loop`) for more controlled, incremental progress.
It is automatically enabled in most of the tests in
`test_indexing_advanced.cpp`.

Major changes are:

- Concrete loop domain: it's either the promotion ID (when
`IdModelEnableOption::Loop` is on) or the concrete ID in `ComputeAtMap`
- Loop index variable: it's either the variable managed by `IdModel`
(when `IdModelEnableOption::Loop` is on) or `ComputeAtMap`



This is part of #2851.