Loop promotion analysis WAR/shortcut and cycle check (#3079)
To workaround the cyclic graph issue (#3072) as much as possible, this
PR skips the full analysis when possible. When skipped, the alternative
analysis is just simply picking the first ID, so it shouldn't have any
constraint with the graph structure. It can be also considered a
shortcut as it would skip the analysis if not necessary.

Also added a check of existence of cycles in a graph as it isn't
supported.