Host irs: add p2p comms (#2970)
# What
Add a new communication IR called `P2PCommunication` representing the
one-sided operations of `send` and `recv`.

# Why

## Used in RS+GEMM Host IR implementation
The Reduce-Scatter+GEMM P2P pipelining has been implemented in ATen in 
#2950. The host IR implementation needs this new IR to express the
send/recv, see #3060

##  Why not using the pre-existing `Communication` class

A `Communication` IR represents in nvFuser a MPI-style collective (e.g.
allgather, reduce-scatter, etc.). For completeness, we also allowed this
class to represent a two-sided `SendRecv` operation, grouping together
the send and recv calls, even though the semantics of `SendRecv` does
not perfectly matches the one of MPI collective. Here is a summary of
the main differences
- `send` and `recv` are one-sided communications, contrarily to
collectives.
- In contrast with other MPI-collectives, the world communicator (i.e.
the `c10d::Backend` initialized with all ranks) can be used to perform
the communication. In contrast, an MPI-collective requires instantiating
a new `c10d::Backend` for each new `Team`.
- `SendRecv` result in two different backend call (i.e.
`c10d::Backend::send` or `c10d::Backend::recv`) depending on the role
sender/receiver. In contrast, all rank participating to a collective
(e.g. `allgather`) call the same primitive (`c10d::Backend::allgather`).
Therefore, having two different IRs for sending and receiving matches
the ATen code better.
- c10d::Backend has different signatures between send/recv on the one
hand and other collectives on the other hand:
    - `send` and `recv` take _one peer rank_ and _one buffer_
    - Collectives take a _team of peer ranks, a src and a dst buffer_.

In addition, for the present use case, we think it is more convenient to
add new IRs, `Send` and `Recv`, rather than using the existing
`SendRecv` IR. Here are the main reasons:

- **Need for a fully symbolic IR**. The `Communication` IR is, for now,
not fully symbolic, since the `root` and `Team` are passed as concrete
values and not symbolic `Val`s. This does not fit the need of The
Reduce-Scatter+GEMM P2P #2950 since, there, the send/recv peer depends
on the index of a symbolic ForLoop.
- Systematically grouping Send and Recv together is not desirable
because we want to be able to typically synchronize differently on those
two calls