add unroll to pointwise scheduler (#3067)
**What is in this PR?** This PR allows unroll on top of vectorization in
pointwise scheduler.
**Why want to do this?** See
[doc](https://docs.google.com/document/d/14GFuGodBatD-g7_Ws--4QEjPZGAAPOPZOhjrETX-gX8/edit#heading=h.mp3o4wf6epgb)
for details.
Unroll may save loaded data when there is outer broadcast Tv in the
fusion. For example `y = gelu(x + bias)`
where x is `T0[i1, i2]`, and bias is `T1[b, i3]`. For this case, 2D
pointwise scheduler is used, when `unroll = 2` in outer dim, we can save
one load of T1 due to the broadcast. For T1 we only need to load one
row:
```
 Array<__half, 8, 8> T18;
  T18.set(__half(0));
  if (b3) {
    loadGlobalToLocal<__half, /*vec_size=*/8, /*is_volatile=*/false, CacheOp::AllLevels>(&T18[0],  &T1[(i0 + i1)]);
  }
```
For T0 we need to load two rows:
```
   #pragma unroll
    for(nvfuser_index_t i6 = 0; i6 < 2; ++i6) {
      loadGlobalToLocal<__half, /*vec_size=*/8, /*is_volatile=*/false, CacheOp::Streaming>(&T17[(8 * i6)],  &T0[(i2 + (T0.logical_size[1LL] * (i6 + nvfuser_zero)))]);
    }
```
**Follow ups:** This PR only enables this feature, but not used in the
heuristics. So, there is no change in the generated code. Further work
is tuning heuristics to allow vectorization & unroll at the same time
for better performance.