Allow linear to take a >2D weight and a >1D bias. (#3073)
As long as the extra dimensions are DID-parallel.

This allows a distributed transformer layer to use linear (instead of
matmul+add) for speed and will simplify the pending #3045.

To avoid ambiguity, this PR also removes the support for 1D weight and
0D bias; otherwise, it's unclear whether a 2D weight is one device
dimension plus a non-device or two non-devices. This support can be
added back by changing the thunder-to-nvFuser bridge to convert a 1D/0D
linear to unsqueeze followed by a 2D/1D linear followed by a squeeze.