Fix fd.ops.linear to compute the right output rank. (#3114)
This PR fixes a bug introduced in #3073. This bug causes
`nvFuser.Tensor` to have a different rank than the corresponding
`TensorView`. This didn't trigger any test failure until I wrote a more
complicated test that `slice`s the output of a linear.

Question for @rdspring1 and/or @kevinstephano: shouldn't this bug be
caught earlier? I guess when the Python frontend finalizes the
definition it should have checked the output `nvFuser.Tensor`s are
consistent with the output `TensorView`s. Wdyt?