Cleanup `applyMmaSwizzleForTMALoad` (#3077)
In the past, we had a `permute_outer_dim` parameter for
`applyMmaSwizzleForTMALoad` because strides in the Mma Op's matrix
descriptor were hard-coded. With hardcoded strides, we had zero
flexibility, and had to use multiple TMA to load smem buffer into a very
specific format. After https://github.com/NVIDIA/Fuser/pull/3002, the
strides were computed based on the schedule, so we do have the
flexibility to choose a memory layout that uses less TMA instructions to
load.