async copy save registers (#2339)
**Issue** : Current CpAsync doesn't allow IfThenElse predicate, becuase
for matmul we need to use inline predicate in cpAsync to ensure out of
boundary positions are filled with 0. However, for non-matmul cases, out
of boundary positions are not used and we can use the IfThenElse
predicate.
**Fix** : Allow using IfThenElse predicate for CpAsync if there is no
matmul in the fusion.
**Influence** : After this change, we can use CpAsync in normalization
kernels with shared memory persistent. It reduces register usage/spills
since it bypass L1/SMEM when copying from gmem to smem.