Add option to disable multidevice at runtime (#3109)
# What
- Add an option to disable setting up multidevice communicator at
runtime through parsing the env `NVFUSER_DISABLE` with the new option
`multidevice`.

- Add missing checks to gracefully skip some recently added tests when
the communicator is unavailable.

# Why

Mainly for debugging. This option quickly disables the multidevice
module even when Fuser is built with multidevice support. We need this
in particular to investigate possible conflicts between nvFuser
communicator and the framework also setting up `c10d::ProcessGroup`