Only use `fence.proxy.async` once (#3136)
According to:
https://docs.nvidia.com/cuda/parallel-thread-execution/#async-proxy,
mbarrier wait already guaranteed correct ordering, so this sync is not
needed:

> The completion of a `cp{.reduce}.async.bulk` operation is followed by
an implicit generic-async proxy fence. So the result of the asynchronous
operation is made visible to the generic proxy as soon as its completion
is observed. Async-group OR mbarrier-based completion mechanism must be
used to wait for the completion of the `cp{.reduce}.async.bulk`
instructions.

Perf after change:
```C++
 Time (%)  Total Time (ns)  Instances  Avg (ns)   Med (ns)   Min (ns)  Max (ns)  StdDev (ns)                                                  Name

 --------  ---------------  ---------  ---------  ---------  --------  --------  -----------  ----------------------------------------------------------------------------------------------------
     23.2          1681558          1  1681558.0  1681558.0   1681558   1681558          0.0  <unnamed>::nvfuser_none_f0_c0_r0_g0(<unnamed>::Tensor<<unnamed>::__half, (int)3, (int)3>, <unnamed>…
     10.8           783356          1   783356.0   783356.0    783356    783356          0.0  nvjet_hsh_192x192_64x3_2x1_v_bz_coopB_NTN
```

Perf after this PR nvFuser/cuBLAS: `46.6%`. Before this PR: `26%` as
shown in https://github.com/NVIDIA/Fuser/issues/3137