Scheduling loop domains (#3107)
Part of #2851 
Closes #3134 

Adds a new scheduler utility function,
`scheduler_utils::scheduleLoopDomainsLike(std::vector<TensorView*> tvs,
std::vector<IterDomain*> reference_loop_domain)`. This works in some
sense in a similar manner as the existing transform propagator but is
more flexible as it does not assume the forward dependency from
root/logical to loop domains. The primary motivation is to simplify
setting up the loop domains of tensors in RoPE-like fusions. See, for
example, the [`SliceSliceConcatConcat`
test](https://github.com/NVIDIA/Fuser/blob/8046359bcc226c613c140fa2762a49fe5c0d7fea/tests/cpp/test_resize.cpp#L4131).
With this PR, all of the manual calls of `setLoopDomain` are eliminated.

While the actual implementation resulted in a little convoluted, the
underlying algorithm is quite simple. When setting up the loop domain of
a given tensor, we first find a path from the reference loop domain to
the root domain of the tensor (not logical domain. See
https://github.com/NVIDIA/Fuser/pull/3107/files#diff-078826ec098dc190fe9aef215a8130a7f5d268dc36f9537b2bb9e6f255d589d6R2939-R2957).
Then, new IDs are created by cloning the reference IDs and then the path
is replayed using the new cloned IDs as the inputs. The new IDs are then
set as the new loop domain of the tensor.

The above algorithm actually has one problem with path finding since the
shortest ValGraph path may not be a valid path within a single
TensorDomain. That's because while a ValGroup can have multiple
definitions, that's still not allowed for IterDomain. To avoid this
problem, I took a two-step traversal, where a backward-only traversal is
first done from the reference domain and then a forward-only traversal
is done to a target tensor. See
https://github.com/NVIDIA/Fuser/pull/3107/files#diff-078826ec098dc190fe9aef215a8130a7f5d268dc36f9537b2bb9e6f255d589d6R2915
for more details.

To do forward or backward only traversals, I added an option to `BFS` to
specify what directions can be taken. See the new test in test_bfs.cpp
for a concrete example case.