A few cleanups (#3146)
- add new IR node `kir::FenceAsyncProxy` and `kir::WgMmaFence`, which
will be lowered as `fence.proxy.async` and `wgmma.fence.sync.aligned` in
`inline_ptx.cpp`. While it is possible to just create the `kir::Asm`
nodes in the passes, I believe it's better to have a separate IR node so
that other passes will not need to worry about details of the PTX
instruction.
- In the past, syncs for TMA store and MMA are inserted in `index.cpp`
and `inline_ptx.cpp`, which is not optimal. I moved the insertion code
to `insert_syncs.cpp`.
- For Hopper MMA whose inputs are not guarded by mbarrier, insert fences
before each mma instruction. These fences were removed in
https://github.com/NVIDIA/Fuser/pull/3139 and
https://github.com/NVIDIA/Fuser/pull/3136, but I believe we should add
them back for this edge case, because there was no mbarrier guaranteeing
the ordering. This is just an edge case, for real kernel, we will only
use the SS variant and operands will always be guarded by mbarrier.

This is just a cleanup, there is no change in the perf of
https://github.com/NVIDIA/Fuser/issues/3137