Remove some Hopper tests (#3160)
There are lots of Hopper matmul tests. These tests comes from
incremental development: every time we add something new feature, we add
a new test for it. While having these tests made a lot of sense in
incremental development, we do not have to keep all of them after things
are developed. For example, if we have:
```
Test A: feature 1
Test B: feature 1 + feature 2
Test C: feature 1 + feature 2 + feature 3
```
While having all the above tests does help us quickly narrow down which
part is wrong if failures happen, we do not have unlimited resources to
run all of them. In this case, we can delete Test B because C already
covers all the features, although C is not as unit as B. I would still
keep A as it is very unit for feature 1.

The latency of Hopper CI is becoming extremely long recently. It is
pretty common to wait >12h, even >24h to get a result. This is
unacceptable. This PR removes some tests while still making sure that
pieces covered by the removed tests are covered somewhere else.