Use parity to wait for mbarrier (#3155)
Today (assuming 4 stages):
```C++
__shared__ uint64_t mbarrier[4];
__shared__ uint32_t tokens[4];
for i in 3:
  tokens[i] = arrive(mbarrier[i]);
  TMA(i);
for i in I:
  tokens[(i + 3) % 4] = arrive(mbarrier[(i + 3) % 4]);
  TMA((i + 3) % 4);
  wait(mbarrier[i % 4], tokens[i % 4]);
  MMA(i);
```

Changed:
```C++
__shared__ uint64_t mbarrier[4];
for i in 3:
  arrive(mbarrier[i]);
  TMA(i);
for i in I:
  arrive(mbarrier[(i + 3) % 4]);
  TMA((i + 3) % 4);
  waitParity(mbarrier[i % 4], (i / 4) % 2);
  MMA(i);
```

Why making this change? The primary reason is to gain more flexibility
on moving code around. In the changed code, the thing to wait is only a
function of loop index, and is computed inline. It can be moved around
by just cut-paste. There is nothing to worry about like "How do I get
which token to wait on?", and there is no need to worry about data races
on `tokens` when moving things around.

Perf together with https://github.com/NVIDIA/Fuser/pull/3153:

```C++
 Time (%)  Total Time (ns)  Instances  Avg (ns)   Med (ns)   Min (ns)  Max (ns)  StdDev (ns)                                                  Name

 --------  ---------------  ---------  ---------  ---------  --------  --------  -----------  ----------------------------------------------------------------------------------------------------
     20.1          1401712          1  1401712.0  1401712.0   1401712   1401712          0.0  <unnamed>::nvfuser_none_f0_c0_r0_g0(<unnamed>::Tensor<<unnamed>::__half, (int)3, (int)3>, <unnamed>…
     11.2           780952          1   780952.0   780952.0    780952    780952          0.0  nvjet_hsh_192x192_64x3_2x1_v_bz_coopB_NTN
```

Perf nvFuser/cuBLAS: `55.7%`, vs only
https://github.com/NVIDIA/Fuser/pull/3153 `52.7%`

The perf diff for the current schedule (64x256 per CTA) is not
significant, but if I use 192x192 tile size for each CTA (same as
cuBLAS, has larger theoretical throughput), the perf improvement is
bigger `1811248 ns` vs `2168390 ns`.