Support using stmatrix to store any shape in registers by tiling to multiple calls to stmatrix for 8x8, 16x8 and 16x16 tiles. (#3131)
This PR adds support to store any 2D piece of memory in registers to
shared memory using stmatrix.
We tile the 2D memory in registers and store it using multiple calls to
stmatrix inside a for-loop.

For example the 2D memory [32, 32] can be tiled to 16 stmatrix calls
(.x1 8x8) or 8 calls (.x2 16x8) or 4 calls (.x4 16x16).
In this PR, we start with a 2D memory say [32, 64], tile it to [(2, 16),
(4, 8) ] -> [4, 2, 16, 8] -> [8, 16, 8] -- That is 8 calls to stmatrix
of type .x2 which stores 16x8 tiles.

For the input [32, 64] once we have transformed to [8, 16, 8], we
swizzle the memory in registers. We then parallelize it such that that
inner-most dims (after merging) is vectorized (with vector length being
the same as the elements being store per thread).


![indexing](https://github.com/user-attachments/assets/cc71dbf3-0f10-4159-ad7b-bfdf298b0a38)



After swizzling and parallelizing we generate indices for the output of
the stmatrix instruction in a hardcoded manner. The indexing logic has
two parts.

1. Computing the index of the tile.

![indexing](https://github.com/user-attachments/assets/7bdaeb33-b4b7-4e6c-b357-dddbc59ce17d)



2. Computing the index inside the tile. The final index is the sum of
the two.

![inner-tile](https://github.com/user-attachments/assets/b623640b-5452-4c08-a728-a89e764d771e)