Remove bit_cast to convert MmaMacroEncode to uint64_t (#3182)
This standardizes how we convert `MmaMacroEncode` to `uint64_t`. Doing
this with a bitcast actually gives us UB because we are not guaranteed
to have the given order for struct fields defined as bitfields. I first
noticed this was happening when using a `libnvfuser.so` compiled as
C++20 with another project linking against that was compiled using
C++17. In that case, the compilation and linking worked properly, but we
wound up with garbage for the `MmaMacro`. I think this PR will address
this kind of issue.