Host IRs: support `IfThenElse` (#3063)
# What
- `HostIrExecutor` support for `IfThenElse`
- `HostIrExecutor` support scalar inputs. This was needed for testing,
were we pass a `bool` as input.

We might wanna move `IfThenElse` and `Predicate` nodes out of the `kir`
namespace, as we previously did for ForLoop #2440. Lmk what you think.

# Why
- It will be needed in the future for adding UCC support in Host IR
RS+GEMM p2p pipelining https://github.com/NVIDIA/Fuser/pull/3060
- Another application would be to add predicate surrounding the compute
and comms in case of Pipeline parallelism. Today, we are simply removing
in each nodes the parts of the graph that should not be executed, thus
resulting in a non SPMD program