Refactor of the circular buffering analysis and allocation pass (#3208)
A refactor of the circular buffering analysis and allocation pass pulled
out from https://github.com/NVIDIA/Fuser/pull/3202. There should be no
behavioral change, just code movement to make it easier to implement
https://github.com/NVIDIA/Fuser/pull/3202.

Changes:
- Let `CircularBufferInfo` track the concrete loop id -> circular
buffered tv mapping. To conveniently do that, `setCircularBufferAxis` is
renamed as `setCircularBufferTv` which not only does the things of the
original `setCircularBufferAxis` but also updates
`circular_buffer_tvs_`, the container used to track the new mapping.
- Move the insertion of the mbarrier allocation, initialization, and
invalidation to `handle(ForLoop* fl)` in the allocation pass.