Rewrite sync insertion for async expressions with commit-wait flavor (`wgmma`, TMA store) (#3191)
This PR changes how we insert async wait expressions on async operations
that have a commit-wait flavor (`wgmma`, TMA store). In the past, we
always just insert a `commit; wait 0;` right after the async operation
itself in the `ReadAfterWriteSyncs` pass, which means, we just force
making these async operations synchronous. With this PR, I have changed
it to:
- `ReadAfterWriteSyncs` will only insert the RAW async wait before the
first read, just like how `__syncthreads();` are inserted.
- Added a new pass `WarAsyncWaitInserter` that runs after the circular
buffering pass that will be responsible for WAR async waits.

Sketch of generated code for the `HopperMatmulTest.HSH_NT_128BSwizzle`
test with `stages=4` and `prefetch=2`:

```C++
// prologue loop:
for i in 0..prefetch :
  TMA load A for stage i;
  TMA load B for stage i;
// main loop:
for i in 0..N-prefetch:
  TMA load A for stage (i + prefetch) % num_stages;
  TMA load B for stage (i + prefetch) % num_stages;
  wait for TMA load of stage i % num_stages;
  wgmma;
  commit wgmma;
  wait wgmma num_stages - prefetch - 1;
// epilogue loop:
for i in N-prefetch..N:
  wait for TMA load of stage i % num_stages;
  wgmma;
commit wgmma;
wait wgmma 0;
store result from register to gmem
```
where the commit and wait of wgmma in the main loop is inserted by the
`WarAsyncWaitInserter` pass, and the top level commit and wait of wgmma
is inserted by the `ReadAfterWriteSyncs` pass.

Note that the `cp.async` is also an async expression of the same flavor.
Instead of unifying it with the new sync insertion approach, I decided
not to touch it to save my time.

Perf of `HopperMatmulTest.HSH_NT_128BSwizzle`:

Looks like pipelining MMA has little perf benefit, but on the other
hand, changing the epilogue from:
```C++
for:
  wait TMA load;
  wgmma;
  commit wgmma;
  wait wgmma 0;
```

into 

```C++
for:
  wait TMA load;
  wgmma;
commit wgmma;
wait wgmma 0;
```

does help the perf (benchmarked with `stages=4`, `prefetch=3`):

```
 Time (%)  Total Time (ns)  Instances  Avg (ns)   Med (ns)   Min (ns)  Max (ns)  StdDev (ns)                                                  Name                                                
 --------  ---------------  ---------  ---------  ---------  --------  --------  -----------  ----------------------------------------------------------------------------------------------------
     14.9          1325753          1  1325753.0  1325753.0   1325753   1325753          0.0  <unnamed>::nvfuser_none_f0_c0_r0_g0(<unnamed>::Tensor<<unnamed>::__half, (int)3, (int)3>, <unnamed>…
      8.8           783100          1   783100.0   783100.0    783100    783100          0.0  nvjet_hsh_192x192_64x3_2x1_v_bz_coopB_NTN   
```

Perf nvFuser/cuBLAS: `59%`, previously `55.7%` according to
https://github.com/NVIDIA/Fuser/issues/3137.